refactor(auth): AuthResponse 제거 및 token 구조 변경
- UserSignupResponse를 SignupResponse로 변경
- SignupResponse와 LoginResponse에 token 필드를 직접 포함

BREAKING CHANGE: token이 더 이상 AuthResponse wrapper에 포함되지 않으며,
각 응답 객체(SignupResponse, LoginResponse)에 직접 포함
