--- src/main/java/com/clover/bookflow/domain/auth/dto/AuthResponse.java
@@ -1,14 +1,0 @@
-package com.clover.bookflow.domain.auth.dto;
-
-import com.clover.bookflow.domain.user.dto.response.UserSignupResponse;
-
-public record AuthResponse(
-    UserSignupResponse userSignupResponse,
-    String token
-) {
-
-  public static AuthResponse from(UserSignupResponse response, String token) {
-    return new AuthResponse(response, token);
-  }
-
-}
