+++ src/main/java/com/clover/bookflow/domain/auth/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.clover.bookflow.domain.auth.dto;
+package com.clover.bookflow.domain.auth.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
