+++ src/main/java/com/clover/bookflow/domain/auth/dto/response/SignupResponse.java
@@ -1,0 +1,17 @@
+package com.clover.bookflow.domain.auth.dto.response;
+
+import com.clover.bookflow.domain.user.entity.User;
+
+public record SignupResponse(
+
+    String email,
+    String nickname,
+    String token
+
+) {
+
+  public static SignupResponse from(User user, String token) {
+    return new SignupResponse(user.getEmail(), user.getNickname(), token);
+  }
+
+}
